NavCompose, pase entre list y screen de pantallas hecho
