+++ app/src/main/java/edu/ucne/roomaplicadaii/presentation/tecnico/TecnicoViewModel.kt
@@ -1,7 +1,5 @@
-package edu.ucne.roomaplicadaii.presentation
+package edu.ucne.roomaplicadaii.presentation.tecnico
 
-import androidx.lifecycle.AbstractSavedStateViewModelFactory
-import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import edu.ucne.roomaplicadaii.data.local.entities.TecnicoEntity
@@ -9,6 +7,7 @@ import edu.ucne.roomaplicadaii.repository.TecnicoRepository
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.stateIn
+import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 
 class TecnicoViewModel(private val repository: TecnicoRepository, private val tecnicoId: Int) :
@@ -23,15 +22,46 @@ class TecnicoViewModel(private val repository: TecnicoRepository, private val te
             started = SharingStarted.WhileSubscribed(5_000),
             initialValue = emptyList()
         )
-
-    fun saveTecnico(tecnico: TecnicoEntity) {
+    fun onSueldoHoraChanged(sueldoHoraStr: String) {
+        val regex = Regex("[0-9]{0,7}\\.?[0-9]{0,2}")
+        if (sueldoHoraStr.matches(regex)) {
+            val sueldoHora = sueldoHoraStr.toDoubleOrNull() ?: 0.0
+            uiState.update {
+                it.copy(
+                    sueldoHora = sueldoHora,
+                )
+            }
+        }
+    }
+    fun onNombresChanged(nombres: String){
+        uiState.update {
+            it.copy(nombres = nombres)
+        }
+    }
+    init {
         viewModelScope.launch {
-            repository.saveTecnico(tecnico)
+            val tecnico = repository.getTecnico(tecnicoId)
+
+            tecnico?.let {
+                uiState.update {
+                    it.copy(
+                        tecnicoId = tecnico.tecnicoId ?: 0,
+                        nombres = tecnico.nombres,
+                        sueldoHora = tecnico.sueldoHora
+                    )
+                }
+            }
+        }
+    }
+
+    fun saveTecnico() {
+        viewModelScope.launch {
+            repository.saveTecnico(uiState.value.toEntity())
         }
     }
 
 
-    companion object {
+    /*companion object {
         fun provideFactory(
             repository: TecnicoRepository
         ): AbstractSavedStateViewModelFactory =
@@ -45,12 +75,8 @@ class TecnicoViewModel(private val repository: TecnicoRepository, private val te
                     return TecnicoViewModel(repository, 0) as T
                 }
             }
-    }
-    fun saveTecnico() {
-        viewModelScope.launch {
-            repository.saveTecnico(uiState.value.toEntity())
-        }
-    }
+    }*/
+
 }
 data class TecnicoUIState(
     val tecnicoId: Int = 0,
