+++ app/src/main/java/edu/ucne/roomaplicadaii/repository/TecnicoRepository.kt
@@ -6,6 +6,9 @@ import edu.ucne.roomaplicadaii.data.local.entities.TecnicoEntity
 class TecnicoRepository(private val tecnicoDao: TecnicoDao) {
 
     suspend fun saveTecnico(tecnico: TecnicoEntity) = tecnicoDao.save(tecnico)
+    suspend fun deleteTecnico(tecnico: TecnicoEntity) = tecnicoDao.delete(tecnico)
 
     fun getTenicos() = tecnicoDao.getAll()
+
+    suspend fun getTecnico(id: Int) = tecnicoDao.find(id)
 }
